Refactor/password-reset-system (#127)
feat(error): implement unified exception handling system

Introduces a comprehensive exception handling system for improved error management, logging, and API responses. This refactoring creates a consistent, hierarchical approach to exception handling across all services.

Core improvements:

Create BankAppBaseException as the root of exception hierarchy
Implement ApiExceptionHandler for centralized error management
Introduce ErrorCode enum for consistent error propagation
Add uniform logging approach with configurable stack traces
Security enhancements:

Improve password reset flow with secure error handling
Add protection against multiple reset attempts
Implement silent handling for sensitive errors
Update token-related exceptions for better reusability
Architecture changes:

Refactor account/user/transaction exceptions to use new hierarchy
Enhance validation error handling with field-level details
Update error status codes for better HTTP semantics
Reorganize shared classes for better maintainability
Developer experience:

Add comprehensive JavaDoc documentation
Improve code formatting and remove unused imports
Replace @value with @Getter and @AllArgsConstructor
Add reverse compatibility for BaseApiError
Testing:

Add comprehensive test coverage for exception handlers
Implement validation error processing tests
Enhance integration tests for security scenarios
This refactoring significantly improves error handling consistency, logging capabilities, and overall system maintainability while enhancing security through better error management.

+++ src/main/java/info/mackiewicz/bankapp/security/repository/PasswordResetTokenRepository.java
@@ -35,7 +35,7 @@ public interface PasswordResetTokenRepository extends JpaRepository<PasswordRese
      * Used for rate limiting - preventing users from requesting too many tokens.
      */
     @Query("SELECT COUNT(t) FROM PasswordResetToken t WHERE t.userEmail = :userEmail AND t.used = false AND t.expiresAt > :now")
-    long countValidTokensByUserEmail(@Param("userEmail") String userEmail, @Param("now") LocalDateTime now);
+    int countValidTokensByUserEmail(@Param("userEmail") String userEmail, @Param("now") LocalDateTime now);
 
     /**
      * Find all active tokens.
